last commit

+++ back_end/app/Http/Controllers/AnalyseController.php
@@ -3,6 +3,7 @@
 namespace App\Http\Controllers;
 
 use App\Models\Analyse;
+use App\Models\ParametreAnalyse;
 use Illuminate\Http\Request;
 use Exception;
 
@@ -12,7 +13,7 @@ class AnalyseController extends Controller
     {
         $analyses = Analyse::all();
         if (is_null($analyses)) {
-            return response()->json(['status' => 404, 'message' => 'Analyses not found']);
+            return response()->json(['status' => 404, 'message' => 'Analyses not found'], 404);
         }
         return response()->json($analyses);
     }
@@ -21,7 +22,7 @@ class AnalyseController extends Controller
     {
         $analyse = Analyse::find($id);
         if (is_null($analyse)) {
-            return response()->json(['status' => 404, 'message' => 'Analyse not found']);
+            return response()->json(['status' => 404, 'message' => 'Analyse not found'], 404);
         }
         return response()->json($analyse);
     }
@@ -30,6 +31,7 @@ class AnalyseController extends Controller
     {
         try {
             $request->validate([
+                'id' => 'required',
                 'codeAnalyse' => 'required|string|unique:analyses',
                 'libelleAnalyse' => 'required|string|unique:analyses',
             ]);
@@ -44,27 +46,36 @@ class AnalyseController extends Controller
     {
         try {
             $request->validate([
-                'codeAnalyse' => 'required|string|unique:analyses',
-                'libelleAnalyse' => 'required|string|unique:analyses',
+                'codeAnalyse' => 'required|string|unique:analyses,codeAnalyse,' . $id,
+                'libelleAnalyse' => 'required|string|unique:analyses,libelleAnalyse,' . $id,
             ]);
             $analyse = Analyse::find($id);
             if (is_null($analyse)) {
-                return response()->json(['status' => 404, 'message' => 'Analyse not found']);
+                return response()->json(['status' => 404, 'message' => 'Analyse not found'], 404);
             }
             $analyse->update($request->all());
             return response()->json(['message' => 'Analyse updated successfully']);
         } catch (Exception $e) {
-            return response()->json(['status' => 'error', 'message' => $e->getMessage()]);
+            return response()->json(['status' => 'error', 'message' => $e->getMessage()], 500);
         }
     }
 
     public function deleteAnalyse($id)
     {
-        $analyse = Analyse::find($id);
-        if (is_null($analyse)) {
-            return response()->json(['status' => 404, 'message' => 'Analyse not found']);
+        try {
+            $analyse = Analyse::find($id);
+            if (is_null($analyse)) {
+                return response()->json(['message' => 'Analyse not found'], 404);
+            }
+            $parametreAnalyse = ParametreAnalyse::where(['codeAnalyse' => $analyse->codeAnalyse])->first();
+            if ($parametreAnalyse) {
+                return response()->json(['status' => 400, 'message' => 'Cette Analyse est liée à une Analyse de Process ou un Parametre d\'analyse et ne peut pas être supprimée.'], 400);
+            } else {
+                $analyse->delete();
+                return response()->json(['message' => 'Analyse deleted successfully']);
+            }
+        } catch (Exception $e) {
+            return response()->json(['status' => 500, 'message' => $e->getMessage()], 500);
         }
-        $analyse->delete();
-        return response()->json(['message' => 'Analyse deleted successfully']);
     }
 }
