+++ back_end/app/Http/Controllers/ParametreAnalyseController.php
@@ -3,6 +3,7 @@
 namespace App\Http\Controllers;
 
 use App\Models\ParametreAnalyse;
+use App\Models\ParametreAnalyseProcess;
 use Exception;
 use Illuminate\Database\Eloquent\ModelNotFoundException;
 use Illuminate\Database\QueryException;
@@ -13,15 +14,15 @@ class ParametreAnalyseController extends Controller
     public function index()
     {
         try {
-            $parametreAnalyses = ParametreAnalyse::with('analyse', 'uniteMesure')->get();
-            if ($parametreAnalyses->isEmpty()) {
+            $parametreAnalyses = ParametreAnalyse::with('analyse', 'uniteMesure')->orderBy('codeAnalyse')->get();
+            if (is_null($parametreAnalyses)) {
                 return response()->json(['status' => 204, 'message' => 'parametre analyse not found']);
             }
             return response()->json($parametreAnalyses);
         } catch (QueryException $e) {
-            return response()->json(['status' => 404, 'message' => $e->getMessage()]);
+            return response()->json(['status' => 404, 'message' => $e->getMessage()], 404);
         } catch (Exception $e) {
-            return response()->json(['status' => 500, 'message' => $e->getMessage()]);
+            return response()->json(['status' => 500, 'message' => $e->getMessage()], 500);
         }
     }
 
@@ -29,6 +30,7 @@ class ParametreAnalyseController extends Controller
     {
         try {
             $request->validate([
+                'id' => 'required',
                 'codeParam' => 'required|string|unique:parametre_analyses',
                 'libelleParam' => 'required|string|unique:parametre_analyses',
                 'codeAnalyse' => 'required|string|exists:analyses,codeAnalyse',
@@ -40,7 +42,7 @@ class ParametreAnalyseController extends Controller
             ])->first();
             if ($existParametreanalyse) {
                 return response()->json([
-                    'error' => 'Parametre d\'analyse déjà existant pour cette analyse et cette unité de mesure'
+                    'message' => 'Parametre d\'analyse déjà existant pour cette analyse et cette unité de mesure'
                 ], 409);
             }
             $parametreAnalyse = ParametreAnalyse::create($request->all());
@@ -49,9 +51,9 @@ class ParametreAnalyseController extends Controller
                 $parametreAnalyse
             ]);
         } catch (QueryException $e) {
-            return response()->json(['status' => 404, 'error' => $e->getMessage()]);
+            return response()->json(['status' => 404, 'message' => $e->getMessage()], 404);
         } catch (Exception $e) {
-            return response()->json(['status' => 500, 'error' => $e->getMessage()]);
+            return response()->json(['status' => 500, 'message' => $e->getMessage()], 500);
         }
     }
 
@@ -61,11 +63,11 @@ class ParametreAnalyseController extends Controller
             $parametreAnalyse = ParametreAnalyse::with('analyse', 'uniteMesure')->findOrFail($id);
             return response()->json($parametreAnalyse);
         } catch (ModelNotFoundException $e) {
-            return response()->json(['error' => 'ParametreAnalyse not found'], 404);
+            return response()->json(['message' => 'ParametreAnalyse not found'], 404);
         } catch (QueryException $e) {
-            return response()->json(['status' => 404, 'message' => $e->getMessage()]);
+            return response()->json(['status' => 404, 'message' => $e->getMessage()], 404);
         } catch (Exception $e) {
-            return response()->json(['status' => 500, 'message' => $e->getMessage()]);
+            return response()->json(['status' => 500, 'message' => $e->getMessage()], 500);
         }
     }
 
@@ -73,14 +75,14 @@ class ParametreAnalyseController extends Controller
     {
         try {
             $request->validate([
-                'codeParam' => 'required|string',
-                'libelleParam' => 'required|string',
+                'codeParam' => 'required|string|unique:parametre_analyses,codeParam,' . $id,
+                'libelleParam' => 'required|string|unique:parametre_analyses,libelleParam,' . $id,
                 'codeAnalyse' => 'required|string|exists:analyses,codeAnalyse',
                 'codeUnite' => 'required|string|exists:unite_mesures,codeUnite',
             ]);
             $parametreAnalyse = ParametreAnalyse::findOrFail($id);
             if (is_null($parametreAnalyse)) {
-                return response()->json(['error' => 'ParametreAnalyse not found'], 404);
+                return response()->json(['message' => 'ParametreAnalyse not found'], 404);
             }
             $existParametreanalyse = ParametreAnalyse::where([
                 'codeAnalyse' => $request->codeAnalyse,
@@ -89,16 +91,16 @@ class ParametreAnalyseController extends Controller
 
             if ($existParametreanalyse) {
                 return response()->json([
-                    'error' => 'Parametre d\'analyse déjà existant pour cette analyse et cette unité de mesure'
+                    'message' => 'Parametre d\'analyse déjà existant pour cette analyse et cette unité de mesure'
                 ], 409);
             }
 
             $parametreAnalyse->update($request->all());
             return response()->json(['message' => 'ParametreAnalyse updated successfully', $parametreAnalyse]);
         } catch (QueryException $e) {
-            return response()->json(['status' => 404, 'error' => $e->getMessage()]);
+            return response()->json(['status' => 404, 'message' => $e->getMessage()], 404);
         } catch (Exception $e) {
-            return response()->json(['status' => 500, 'error' => $e->getMessage()]);
+            return response()->json(['status' => 500, 'message' => $e->getMessage()], 500);
         }
     }
 
@@ -106,14 +108,18 @@ class ParametreAnalyseController extends Controller
     {
         try {
             $parametreAnalyse = ParametreAnalyse::findOrFail($id);
+            $parametreAnalyseProcess = ParametreAnalyseProcess::where(['codeParam' => $parametreAnalyse->codeParam])->first();
+            if ($parametreAnalyseProcess) {
+                return response()->json(['status' => 400, 'message' => 'Cette parametre d\'Analyse est liée à une Analyse de Process et ne peut pas être supprimée.'], 400);
+            }
             $parametreAnalyse->delete();
             return response()->json(['message' => 'ParametreAnalyse deleted successfully']);
         } catch (ModelNotFoundException $e) {
-            return response()->json(['error' => 'ParametreAnalyse not found'], 404);
+            return response()->json(['message' => 'ParametreAnalyse not found'], 404);
         } catch (QueryException $e) {
-            return response()->json(['status' => 404, 'error' => $e->getMessage()]);
+            return response()->json(['status' => 404, 'message' => $e->getMessage()], 404);
         } catch (Exception $e) {
-            return response()->json(['status' => 500, 'error' => $e->getMessage()]);
+            return response()->json(['status' => 500, 'message' => $e->getMessage()], 500);
         }
     }
 }
