+++ back_end/app/Http/Controllers/PlanControleController.php
@@ -3,6 +3,7 @@
 namespace App\Http\Controllers;
 
 use App\Models\PlanControle;
+use App\Models\ProcessFabrication;
 use App\Models\Produit;
 use Exception;
 use Illuminate\Database\Eloquent\ModelNotFoundException;
@@ -31,7 +32,7 @@ class PlanControleController extends Controller
             $existPlanControle = PlanControle::where(['codeProduit' => $request->codeProduit])->first();
             if ($existPlanControle) {
                 return response()->json([
-                    'error' => 'Un plan de contrôle existe déjà pour ce produit'
+                    'message' => 'Un plan de contrôle existe déjà pour ce produit'
                 ], 409);
             }
             PlanControle::create($request->all());
@@ -47,8 +48,8 @@ class PlanControleController extends Controller
     public function showPlanControle($id)
     {
         try {
-            $planControle = PlanControle::with('produit')->findOrFail($id);
-            return response()->json(['planControle' => $planControle]);
+            $planControle = PlanControle::with('produit')->where(['codePlan' => $id])->first();
+            return response()->json($planControle);
         } catch (ModelNotFoundException $e) {
             return response()->json(['status' => 'error', 'message' => 'PlanControle not found'], 404);
         } catch (Exception $e) {
@@ -61,11 +62,17 @@ class PlanControleController extends Controller
         try {
 
             $request->validate([
-                'codePlan' => 'required|string|unique:plan_controles',
+                'codePlan' => 'required|string|unique:plan_controles,codeplan,' . $id,
                 'etatPlan' => 'required|string',
                 'codeProduit' => 'required|string|exists:produits,codeProduit',
             ]);
             $planControle = PlanControle::findOrFail($id);
+            // $existPlanControle = PlanControle::where(['codeProduit' => $request->codeProduit])->first();
+            // if ($existPlanControle) {
+            //     return response()->json([
+            //         'message' => 'Un plan de contrôle existe déjà pour ce produit'
+            //     ], 409);
+            // }
             $planControle->update($request->all());
             return response()->json(['message' => 'plan controle a ete modifier', 'planControle' => $planControle]);
         } catch (QueryException $e) {
@@ -77,12 +84,24 @@ class PlanControleController extends Controller
 
     public function deletePlanControle($id)
     {
-        $planControle = PlanControle::findOrFail($id);
-        if (is_null($planControle)) {
-            return response()->json(['status' => 404, 'message' => 'PlanControle not found']);
+        try {
+            $planControle = PlanControle::findOrFail($id);
+            $ProcessFabrication = ProcessFabrication::where(['codePlan' => $planControle->codePlan])->first();
+            if ($ProcessFabrication) {
+                return response()->json(['status' => 500, 'message' => 'Cette Plan de Controle a un process de fabrication et ne peut pas être supprimée.'], 500);
+            }
+            if (is_null($planControle)) {
+                return response()->json(['status' => 404, 'message' => 'PlanControle not found'], 404);
+            }
+            $planControle->delete();
+            return response()->json(['message' => 'PlanControle deleted successfully']);
+        } catch (ModelNotFoundException $e) {
+            return response()->json(['message' => 'ParametreAnalyse not found'], 404);
+        } catch (QueryException $e) {
+            return response()->json(['status' => 404, 'message' => $e->getMessage()], 404);
+        } catch (Exception $e) {
+            return response()->json(['status' => 500, 'message' => $e->getMessage()], 500);
         }
-        $planControle->delete();
-        return response()->json(['message' => 'PlanControle deleted successfully']);
     }
 }
