+++ back_end/app/Http/Controllers/ProduitController.php
@@ -2,6 +2,7 @@
 
 namespace App\Http\Controllers;
 
+use App\Models\PlanControle;
 use App\Models\Produit;
 use Exception;
 use Illuminate\Database\QueryException;
@@ -13,7 +14,7 @@ class ProduitController extends Controller
     {
         $produits = Produit::all();
         if (is_null($produits)) {
-            return response()->json(['status' => 404, 'message' => 'Produits not found']);
+            return response()->json(['status' => 404, 'message' => 'Produits not found'], 404);
         }
         return response()->json($produits);
     }
@@ -31,6 +32,7 @@ class ProduitController extends Controller
     {
         try {
             $request->validate([
+                'id' => 'required',
                 'codeProduit' => 'required|string|unique:produits',
                 'libelleProduit' => 'required|string|unique:produits',
             ]);
@@ -53,7 +55,7 @@ class ProduitController extends Controller
                 return response()->json(['status' => 404, 'message' => 'Produit not found']);
             }
             $produit->update($request->all());
-            return response()->json($produit); // Return the updated product
+            return response()->json($produit);
         } catch (QueryException $e) {
             return response()->json(['status' => 'error', 'message' => $e->getMessage()], 500);
         } catch (Exception $e) {
@@ -62,13 +64,17 @@ class ProduitController extends Controller
     }
 
 
-    public function deleteproduit($id)
+    public function deleteProduit($id)
     {
         $produit = Produit::find($id);
         if (is_null($produit)) {
-            return response()->json(['status' => 404, 'message' => 'Produit not found']);
+            return response()->json(['status' => 404, 'message' => 'Produit not found'], 404);
+        }
+        $planControle = PlanControle::where(['codeProduit' => $produit->codeProduit])->first();
+        if ($planControle) {
+            return response()->json(['status' => 400, 'message' => 'Impossible de supprimer ce produit car il est utilisé dans un plan de contrôle'], 400);
         }
         $produit->delete();
-        return response()->json(['message' => 'Produit deleted successfully']);
+        return response()->json(['message' => 'Produit deleted successfully'], 200);
     }
 }
